server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pillow>=10.0.0
aiofiles>=23.0
pyyaml>=6.0
pydantic>=2.0
python-multipart>=0.0.6
zeroconf>=0.131.0
